Store files when running Actions in server
